Fix trustline deletion, improve trustline handling
Make sure rippling flag is off when deleting a trustline
(set limit=0)

Trustline currencies < 3 chars are considered "no currency"
when setting a trustline. Allows us to delete ones we set
by mistake.

If an account shows a balance for a trustline we have not
seen before, add it to the trustline list and treat it like
any other.