Redesign data API to handle dynamic handling of data display
Add ohlc_dataset_view to hold a dataset + resampled series and
leave only hdf5 code in the data manager

Use smart_ptrs to manage data manager and views and other objects
that must be shared and have variable lifetimes

Rework GUI to fetch data/network options from the new places and.
move some logic into the exchange as data is subscribed to on a per
exchange basis.

Many tweaks to handle the reworking of graph ownership and creation