Remove candlestick update code from mainwindow into bitstamp exchange
More modularization of the candle/data management and updates of plots.

Data is now completely managed by the data manager and views, plots
and updates are independent of the mainwindow, with each exchange object
managing its own timer and update mechanism.

Lots of function cleanup, renaming and consolidation of utilities