Add RSI - Allow indicators to create new graph and align with main price plot
Indicators that do not output data in the same time/price axes as the
main price plot need to go into new graphs instead of appearing as
overlays.

Fix the old code to allow an extra plot (needs to be extended to support
many) to appear under the main one. Fix the yRight scale/margin to match
the main plot to keep everything aligned