Fix websocket stream connect/disconnect
Each ticker maintains a map of streams and each is un/subscribed
individually. Previously there was (wrongly) one websocket per ticker
which caused trade to overwrite orders etc.

Bind handlers to currency pairs so that no lookups need to be done
during the handler process (costly decoding of channel data)

Clean up debug messages and save/restore of state.

GUI could be improved for the case when many tickers are available and
subscribed to.