Fix live data volume and series updates
Volumes were being added/appended incorrectly due to the way that
missing data was checked/added when gaps appear in live trades.
We use time stamps to aappend and update instead of indexes which
simplifies the function and fixes the error

Also switch to using time stamps and clearing live data using remove_if
instead of index checks to fix an issue with last data display.