Move ownership of plot curves into plotting module
The Data module owns the plot data, and used to own the curves too, but
this places a dependency on the plotting infrastructure which we do not
want. Moving all curve ownership into plot/ makes a cleaner separation
of the code. The both still use qt/qwt for the ohlc data types and
qvector storage, but the plotting module now depends on data and unit
tests for data do not pull in any plot based code.

Transfer ownership of live data plot into plotting module and ensure
it behaves as previously. Todo : needs to support live data of any
candle resolution rather than only the minute candles

Clean up use of #includes throughout all places affected by this
change.