Improve indicator plot speed, create plot lib
Create grox_plot subdir/lib and tidy #includes

Create new timebased_data_curve type that overrides
the DrawSeries function to reduce plotting from the entire
series, to just the visible section. This is the same as is already done
in the ohlc plot code, but now the timebased_data abstraction handles it
for both cases.