Replace all https/websocket calls in grox with qt versions
Remove https and websocket code using boost::asio except for tests which
are useful for reference. Move asio code into testing subdir so that it
can still be referred to  ut is not compiled into main app.

Qt-networking calls are still not completing due to probable issue with
completion handlers not being triggered in grox itself (tests work fine)