xrpl Qt network, Remove currency_type, update all currency based apis/functions
Adding support for the new Qt based networking into the xrpl_network
exchange classes required updating the API to handle ticker subscription
based on ticker names using currency pairs. A lot of code relied on
xrp/usd/etc currency "type" which is not scalable when multiple
exchanges/tokens etc are handled (bitstamp/gatehub, trustlines for
example).

Removed the currency_type class and reworked all the functions that used
it to instead query the currency code/issuer directly.

An extension will be to support the exchange name in the currency so that
the same tokens on different networks can be easily handled.