Support p2300 stdexec Senders for async management of Qt/Pika threads
This is a first implementation of stdexec/pika senders with the new
Qt based network handling. Network requests are made on the Qt main thread
and the response is transferred to pika where json decoding and handling
of internal structures is done. Transfer back to Qt for more network
requests of for GUI updates.

So for only the initial bitstamp initialization has been converted to
use stdexec