New function to get first available ticker data from Bitstamp
When a new ticker is subscribed to, data must be downloaded but we don't
know how far back in history data exists for, query bitstamps internal
price history api to get a good guess of when we can fetch data from.

Add new sender async access and upgrade the OHLC candlestick update
mechanism to be sender based and add rudimentry exit/error handling to
exit the sender chain

Rename a bunch of sender based api functions to clarify usage.