Overhaul orderbook event handling with senders and modularize exchange dir
When tickers are subscribed to, they bind callbacks for orderbook
handling and price (trade) updates. This removes a dependecy on the
exchange to issue callbacks and insteead places it on the tickers
themselves which can be subscribed to.

Add sender handling and Qt::InvokeMethod calls to ensure that procesing
is done on pika threads and gui updates are handled on Qt threads