Simplify/unify handling of stream subscriptions + load/save qsettings
price_data is not a websocket datastream, but we include it in the
streams list and handle it like any other stream when registering.

We do not need to store tickers separately in the ini/settings as the
stream subscriptions include this information already.

The creation of the https requests and timer updates of price data is
still handled separately, but this can now be moved into the
subscription handling for a cleaner setup.