Redesign ticker subscription interface/API
Asynchronous loading of networks and subscription information causes
ordering of gui creation to be a problem. Use a model/view design for a
ticker subscription tree to simplify the gui and move the setting of
checked items and subscription into the gui create function instead of
using a separate on in the gui mainwidow startup.

Still some wrinkles with unsubscribing/resubscribing that need to be
cleared up and the layout of saved window states need to be fixed as the
loading of setting and creating of gui controls is now not well ordered.