Update xrpl orderbook handling code to latest api
Senders were added for bitstamp/exchange and new api methods for
handling tickers/streams were added - this brings the xrp ledger support
for orderbook processing up to the same level.

Fix some issues with the way json was parsed and handled that make it
less dependent on key/value ordering in the json orderbook.