Bring XRPL orderbook websocket code in line with latest senders
Update the xrpl subscription to use sender based websocket and fix the
generation of streams and subscription to use the new pub/sub style
registration and hooks that the bitstamp code uses.

Clean up orderbook json handling and move xrpl orderbook plot into its
own dockwindow. This breaks the ability to overlay multiple orderbooks
on the same graph and needs to be reworked to enable it again.