Redesign indicators api/construction to remove need for type in all headers
Previously, all types had to be declared and functions to instantiate
and execute the indicators/algorithms had to kniow the type, this made
addition of new arbitrage type algorithms using the same API difficult.

Using a class heirarchy removes some flexibility in arbitrary
definitions, but adds more support for adding new types - especially
plugins - if they are needed later, since the API uses some virtual
calls to execute the algoriths,

The main loops are performed without internal virtual calls, but the
function to execute them is virtual on the outside.

Lots of small changes to the way indicators are stored, passed and
arranged internnaly for widget/gui creation.

Registration of pubsub for incomoing data is not yet re-implemented.