Adjust shutdown to try to prevent hangs by closing networks asynchronously
The network uses Qt networking/threads and so needs the application
messaging loop to be running before it has finished closing all
connections. When a closeevent happens, we spawn a task to shut down the
networks and ignore the closeevent, then when the network shutdown has
completed, we close the application.