Overhaul of currency exchange algorithm/widget prior to using new control
This gets most of the functionaly we want running, but lacks a simple
way of binding the algorithm/indicator to a given wallet/network/account
etc - the gui works ok, but we need an improved control setu pt convert
indicator params to controls that allows customization of parameters
more simply.

Algorithms/indicators can be registered with an exchange wallet to bind
functionality of algorithms with features supported by a particular
exchange.

Next step is to integrate the new widget_control framework with the
existing indicator gui generation.