Allow indicator algorithm (trade) to run in partitioned chunks
If a back-testing algorithm is run on an entire daataset, it may take
a significant time, during this timer, new data is arriving and should
be appended to the dataset that is being used for calculations.

We add a simple partitioning mechanism that allows a long running
algorithm to operate in chunks and take/release the lock for each
chiunk, allowing new dat to be appended during the breaks.

Iterators are recomputed at the start off each chunk to ensure that
invalidated pointers do not cause segfaults.

Currently, only buy/sell algoriithms implement this, but it can be
extended to all algorithms.

Known bbug is that if new data arrives and triggers an algorithm while
the algorithm is running, some erronreous data may be inserted in the
result. This need to be fixed by a different indexing schem for
algorithm startin (ie use the known endpoint, not jjust the last N)